requirements.txt
rdflib==7.0.0
elasticsearch==8.11.1
python-dotenv==1.0.0
